+++ app/Http/Controllers/ProductController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Product;
 use Illuminate\Http\Request;
 use CloudinaryLabs\CloudinaryLaravel\Facades\Cloudinary;
+use Illuminate\Support\Facades\Auth;
 
 class ProductController extends Controller
 {
@@ -23,36 +24,40 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
-        try {
-            $request->validate([
-                'name' => 'required|string|max:255',
-                'description' => 'nullable|string',
-                'price' => 'required|numeric',
-                'stock' => 'required|numeric',
-                'image_url' => 'required|image',
-            ]);
+        if (Auth::user()) {
+            try {
+                $request->validate([
+                    'name' => 'required|string|max:255',
+                    'description' => 'nullable|string',
+                    'price' => 'required|numeric',
+                    'stock' => 'required|numeric',
+                    'image_url' => 'required|image',
+                ]);
 
-            $file = $request->file('image_url');
-            $cloudinaryUpload = Cloudinary::upload($file->getRealPath(), ['folder' => 'sonrisa']);
+                $file = $request->file('image_url');
+                $cloudinaryUpload = Cloudinary::upload($file->getRealPath(), ['folder' => 'sonrisa']);
 
-            $public_id = $cloudinaryUpload->getPublicId();
-            $url = $cloudinaryUpload->getSecurePath();
+                $public_id = $cloudinaryUpload->getPublicId();
+                $url = $cloudinaryUpload->getSecurePath();
 
-            $product = Product::create([
-                "name" => $request->name,
-                "description" => $request->description,
-                "price" => $request->price,
-                "stock" => $request->stock,
-                "image_url" => $url,
-                "public_id" => $public_id
-            ]);
+                $product = Product::create([
+                    "name" => $request->name,
+                    "description" => $request->description,
+                    "price" => $request->price,
+                    "stock" => $request->stock,
+                    "image_url" => $url,
+                    "public_id" => $public_id
+                ]);
 
-            return response()->json([
-                'message' => 'El producto se ha creado correctamente',
-                'product' => $product,
-            ], 201);
-        } catch (\Exception $e) {
-            return response()->json(['error' => $e->getMessage()], 500);
+                return response()->json([
+                    'message' => 'El producto se ha creado correctamente',
+                    'product' => $product,
+                ], 201);
+            } catch (\Exception $e) {
+                return response()->json(['error' => $e->getMessage()], 500);
+            }
+        } else {
+            return response()->json(['error' => 'Usuario no autorizado'], 401);
         }
     }
 
@@ -79,41 +84,45 @@ class ProductController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        try {
-            $product = Product::findOrFail($id);
+        if (Auth::user()) {
+            try {
+                $product = Product::findOrFail($id);
 
-            $url = $product->image_url;
-            $public_id = $product->public_id;
+                $url = $product->image_url;
+                $public_id = $product->public_id;
 
-            $request->validate([
-                'name' => 'required|string|max:255',
-                'description' => 'nullable|string',
-                'price' => 'required|numeric',
-                'stock' => 'required|numeric',
-                'image_url' => 'required|image',
-            ]);
+                $request->validate([
+                    'name' => 'required|string|max:255',
+                    'description' => 'nullable|string',
+                    'price' => 'required|numeric',
+                    'stock' => 'required|numeric',
+                    'image_url' => 'required|image',
+                ]);
 
-            if ($request->hasFile('image_url')) {
-                Cloudinary::destroy($public_id);
-                $file = $request->file('image_url');
-                $cloudinaryUpload = Cloudinary::upload($file->getRealPath(), ['folder' => 'sonrisa']);
+                if ($request->hasFile('image_url')) {
+                    Cloudinary::destroy($public_id);
+                    $file = $request->file('image_url');
+                    $cloudinaryUpload = Cloudinary::upload($file->getRealPath(), ['folder' => 'sonrisa']);
 
-                $url = $cloudinaryUpload->getSecurePath();
-                $public_id = $cloudinaryUpload->getPublicId();
+                    $url = $cloudinaryUpload->getSecurePath();
+                    $public_id = $cloudinaryUpload->getPublicId();
+                }
+
+                $product->update([
+                    "name" => $request->name,
+                    "description" => $request->description,
+                    "price" => $request->price,
+                    "stock" => $request->stock,
+                    "image_url" => $url,
+                    "public_id" => $public_id
+                ]);
+
+                return response()->json(['message' => 'Producto actualizado correctamente', $product], 200);
+            } catch (\Exception $e) {
+                return response()->json(['error' => $e->getMessage()], 500);
             }
-
-            $product->update([
-                "name" => $request->name,
-                "description" => $request->description,
-                "price" => $request->price,
-                "stock" => $request->stock,
-                "image_url" => $url,
-                "public_id" => $public_id
-            ]);
-
-            return response()->json(['message' => 'Producto actualizado correctamente', $product], 200);
-        } catch (\Exception $e) {
-            return response()->json(['error' => $e->getMessage()], 500);
+        } else {
+            return response()->json(['error' => 'Usuario no autorizado'], 401);
         }
     }
 
@@ -122,18 +131,22 @@ class ProductController extends Controller
      */
     public function destroy(string $id)
     {
-        try {
-            $product = Product::findOrFail($id);
+        if (Auth::user()) {
+            try {
+                $product = Product::findOrFail($id);
 
-            if (!$product) {
-                return response()->json(['message' => 'Producto no encontrado'], 404);
+                if (!$product) {
+                    return response()->json(['message' => 'Producto no encontrado'], 404);
+                }
+
+                $product->delete();
+
+                return response()->json(['message' => 'Producto eliminado correctamente'], 200);
+            } catch (\Exception $e) {
+                return response()->json(['error' => $e->getMessage()], 500);
             }
-
-            $product->delete();
-
-            return response()->json(['message' => 'Producto eliminado correctamente'], 200);
-        } catch (\Exception $e) {
-            return response()->json(['error' => $e->getMessage()], 500);
+        } else {
+            return response()->json(['error' => 'Usuario no autorizado'], 401);
         }
     }
 }
